Add a ResourceTag nested object and adapt API to new UserTag model accordingly
