+++ src/backend/interfaces/REST/py.server/caliopen_api/user/tag.py
@@ -20,7 +20,7 @@ log = logging.getLogger(__name__)
 
 
 @resource(collection_path='/tags',
-          path='/tags/{name}',
+          path='/tags/{tag_id}',
           name='Tag',
           factory=DefaultContext
           )
@@ -35,13 +35,13 @@ class TagAPI(Api):
     @view(renderer='json', permission='authenticated')
     def collection_post(self):
         """Create a new tag for an user."""
-        params = self.request.swagger_data
+        params = self.request.swagger_data['tag']
         user = CoreUser.get(self.user.user_id)
         user_tags = user.tags
         if params['name'] in [x.name for x in user_tags]:
             raise HTTPConflict('Tag {} already exit'.format(params['name']))
         tag = CoreTag.create(self.user, name=params['name'])
-        tag_url = self.request.route_path('Tag', name=tag.name)
+        tag_url = self.request.route_path('Tag', tag_id=tag.tag_id)
         self.request.response.location = tag_url.encode('utf-8')
         # XXX return a Location to get tag not send it direct
         return {'location': tag_url}
@@ -49,20 +49,19 @@ class TagAPI(Api):
     @view(renderer='json', permission='authenticated')
     def collection_get(self):
         """Return list of system and user tags."""
-        sys_tags = Configuration('global').get('system.default_tags')
         user = CoreUser.get(self.user.user_id)
         user_tags = user.tags
         tags = []
-        for tag in sys_tags:
-            tags.append({'name': tag['name'], 'type': 'system'})
         for tag in user_tags:
-            tags.append({'name': tag.name, 'type': 'user'})
+            tags.append({'tag_id': tag.tag_id,
+                         'name': tag.name,
+                         'type': tag.type})
         return {'tags': tags, 'count': len(tags)}
 
     @view(renderer='json', permission='authenticated')
     def get(self):
         """Get a simgle user tag."""
-        name = self.request.swagger_data["name"]
+        name = self.request.swagger_data["tag_id"]
         tag = UserTagObject(self.user.user_id, name=name)
         tag.get_db()
         tag.unmarshall_db()
@@ -71,8 +70,8 @@ class TagAPI(Api):
     @view(renderer='json', permission='authenticated')
     def delete(self):
         """Delete an user tag."""
-        name = self.request.swagger_data["name"]
-        tag = UserTagObject(self.user.user_id, name=name)
+        tag_id = self.request.swagger_data["tag_id"]
+        tag = UserTagObject(self.user.user_id, tag_id=tag_id)
         try:
             tag.get_db()
         except NotFound:
